+++ ChangeLog.md
@@ -1,5 +1,9 @@
 # Changelog for check_cgminer
 
+## 0.2.1.0
+
+- Parse hash rate empty string as zero
+
 ## 0.2.0.0
 
 - Add support for fan speed monitoring